// planner-web/src/api/generated/model/createRecurringExpenseDtoCustomIntervalUnit.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Financial Planner API
 * Versioned API for normalized financial plans
 * OpenAPI spec version: 1.0
 */

export type CreateRecurringExpenseDtoCustomIntervalUnit =
  (typeof CreateRecurringExpenseDtoCustomIntervalUnit)[keyof typeof CreateRecurringExpenseDtoCustomIntervalUnit]

export const CreateRecurringExpenseDtoCustomIntervalUnit = {
  week: "week",
  month: "month",
} as const